Debugger.

+++ aliyun-java-sdk-iot/src/main/java/com/aliyuncs/iot/model/v20190730/GetPoolFunctionsByIdListForTmallGenieResponse.java
@@ -311,7 +311,7 @@ public class GetPoolFunctionsByIdListForTmallGenieResponse extends AcsResponse {
 
 			private String name;
 
-			private Long paraOrder;
+			private Integer paraOrder;
 
 			private String direction;
 
@@ -343,11 +343,11 @@ public class GetPoolFunctionsByIdListForTmallGenieResponse extends AcsResponse {
 				this.name = name;
 			}
 
-			public Long getParaOrder() {
+			public Integer getParaOrder() {
 				return this.paraOrder;
 			}
 
-			public void setParaOrder(Long paraOrder) {
+			public void setParaOrder(Integer paraOrder) {
 				this.paraOrder = paraOrder;
 			}
 
@@ -382,7 +382,7 @@ public class GetPoolFunctionsByIdListForTmallGenieResponse extends AcsResponse {
 
 			private String identifier;
 
-			private Long paraOrder;
+			private Integer paraOrder;
 
 			private String direction;
 
@@ -408,11 +408,11 @@ public class GetPoolFunctionsByIdListForTmallGenieResponse extends AcsResponse {
 				this.identifier = identifier;
 			}
 
-			public Long getParaOrder() {
+			public Integer getParaOrder() {
 				return this.paraOrder;
 			}
 
-			public void setParaOrder(Long paraOrder) {
+			public void setParaOrder(Integer paraOrder) {
 				this.paraOrder = paraOrder;
 			}
 
@@ -553,7 +553,7 @@ public class GetPoolFunctionsByIdListForTmallGenieResponse extends AcsResponse {
 
 			private String name;
 
-			private Long paraOrder;
+			private Integer paraOrder;
 
 			private String dataSpecs;
 
@@ -583,11 +583,11 @@ public class GetPoolFunctionsByIdListForTmallGenieResponse extends AcsResponse {
 				this.name = name;
 			}
 
-			public Long getParaOrder() {
+			public Integer getParaOrder() {
 				return this.paraOrder;
 			}
 
-			public void setParaOrder(Long paraOrder) {
+			public void setParaOrder(Integer paraOrder) {
 				this.paraOrder = paraOrder;
 			}
